pyproject.toml
[project]
name = "rptgen"
version = "0.1.0"
description = "Utility to help me generate reports"
authors = [
    {name = "Tony Gilkerson",email = "tonygilkerson@yahoo.com"}
]
readme = "README.md"
requires-python = ">3.9"
dependencies = [
    "pyyaml>=6.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "src" },
]

[tool.poetry.scripts]
rptgen = "src.main:main"
